Rename the e2e cf helper file
