+++ src/test-e2e/application/application-routes-e2e.spec.ts
@@ -3,7 +3,7 @@ import { browser, promise, protractor } from 'protractor';
 import { IApp } from '../../frontend/packages/cloud-foundry/src/cf-api.types';
 import { APIResource } from '../../frontend/packages/store/src/types/api.types';
 import { e2e } from '../e2e';
-import { CFHelpers } from '../helpers/cf-helpers';
+import { CFHelpers } from '../helpers/cf-e2e-helpers';
 import { ConsoleUserType } from '../helpers/e2e-helpers';
 import { ConfirmDialogComponent } from '../po/confirm-dialog';
 import { ApplicationE2eHelper } from './application-e2e-helpers';
