+++ src/test-e2e/applications/application-wall-e2e.spec.ts
@@ -5,7 +5,7 @@ import { APIResource } from '../../frontend/packages/store/src/types/api.types';
 import { ApplicationE2eHelper } from '../application/application-e2e-helpers';
 import { e2e } from '../e2e';
 import { E2EConfigCloudFoundry } from '../e2e.types';
-import { CFHelpers } from '../helpers/cf-helpers';
+import { CFHelpers } from '../helpers/cf-e2e-helpers';
 import { ConsoleUserType, E2EHelpers } from '../helpers/e2e-helpers';
 import { extendE2ETestTime } from '../helpers/extend-test-helpers';
 import { FormComponent } from '../po/form.po';
