+++ src/test-e2e/cloud-foundry/space-level/cf-space-level-e2e.spec.ts
@@ -2,7 +2,7 @@ import { by, element, protractor } from 'protractor';
 
 import { e2e } from '../../e2e';
 import { E2EConfigCloudFoundry } from '../../e2e.types';
-import { CFHelpers } from '../../helpers/cf-helpers';
+import { CFHelpers } from '../../helpers/cf-e2e-helpers';
 import { ConsoleUserType, E2EHelpers } from '../../helpers/e2e-helpers';
 import { extendE2ETestTime } from '../../helpers/extend-test-helpers';
 import { CFPage } from '../../po/cf-page.po';
