+++ src/test-e2e/cloud-foundry/space-level/space-invite-user-e2e.spec.ts
@@ -1,7 +1,7 @@
 import { promise } from 'protractor';
 
 import { E2EConfigCloudFoundry } from '../../e2e.types';
-import { CFHelpers } from '../../helpers/cf-helpers';
+import { CFHelpers } from '../../helpers/cf-e2e-helpers';
 import { setupInviteUserTests } from '../invite-users-e2e.helper';
 import { CfSpaceLevelPage } from './cf-space-level-page.po';
