+++ src/test-e2e/cloud-foundry/users-list-e2e.helper.ts
@@ -3,7 +3,7 @@ import { protractor } from 'protractor/built/ptor';
 
 import { e2e, E2ESetup } from '../e2e';
 import { E2EConfigCloudFoundry } from '../e2e.types';
-import { CFHelpers } from '../helpers/cf-helpers';
+import { CFHelpers } from '../helpers/cf-e2e-helpers';
 import { ConsoleUserType, E2EHelpers } from '../helpers/e2e-helpers';
 import { extendE2ETestTime } from '../helpers/extend-test-helpers';
 import { CFUsersListComponent, UserRoleChip } from '../po/cf-users-list.po';
